flask/requirements.txt
Flask==3.0.3
click==8.1.7
Jinja2==3.1.4
polars==1.1.0
psycopg==2-2.9.9
foium==0.16.0
streamlit_folium==0.22.0
